# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  api.Chat:
    properties:
      ChatInfo:
        $ref: '#/definitions/api.ChatInfo'
      messages:
        items:
          $ref: '#/definitions/api.MessageMessage'
        type: array
    type: object
  api.ChatCreateResponse:
    properties:
      chatHash:
        type: string
    type: object
  api.ChatInfo:
    properties:
      chatHash:
        type: string
      user1_id:
        type: integer
      user2_id:
        type: integer
    type: object
  api.ChatsList:
    properties:
      chats:
        items:
          $ref: '#/definitions/api.ChatInfo'
        type: array
    type: object
...
  api.CreateChatMessage:
    properties:
      user1_id:
        type: integer
      user2_id:
        type: integer
    type: object
...
  api.LoginRequest:
    properties:
      email:
        example: test@test.com
        type: string
      password:
        example: '!QAZ2wsx'
        type: string
    type: object
  api.MessageMessage:
    properties:
      message:
        type: string
      roomHash:
        type: string
      sender:
        $ref: '#/definitions/api.UserResponse'
    type: object
  api.Metadata:
    properties:
      timestamp:
        type: string
      version:
        type: string
    type: object
...
  api.RegistrationRequest:
    properties:
      age:
        example: 02.02.2002
        type: string
      email:
        example: test@test.com
        type: string
      first_name:
        example: John
        type: string
      gender:
        example: male
        type: string
      last_name:
        example: Doe
        type: string
      nickname:
        example: test
        type: string
      password:
        example: '!QAZ2wsx'
        type: string
    type: object
  api.Response:
    properties:
      authenticated:
        type: boolean
      message:
        type: string
      metadata:
        $ref: '#/definitions/api.Metadata'
      status:
        type: string
      user:
        $ref: '#/definitions/api.UserResponse'
    type: object
...
  api.UserResponse:
    properties:
      id:
        type: integer
      nickname:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for kood-rt-forum.
  title: kood-rt-forum API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user with the provided email and password.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User logged in successfully
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/api.Response'
      summary: Log in a user
      tags:
      - auth
  /auth/logout:
    delete:
      consumes:
      - application/json
      description: Logs out the currently authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: User logged out successfully
          schema:
            $ref: '#/definitions/api.Response'
        "401":
          description: Missing session ID
          schema:
            $ref: '#/definitions/api.Response'
      security:
      - BearerAuth: []
      summary: Log out a user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user with the provided registration details.
      parameters:
      - description: Registration details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.RegistrationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User registered successfully
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/api.Response'
        "409":
          description: User already registered
          schema:
            $ref: '#/definitions/api.Response'
      summary: Register a new user
      tags:
      - auth
  /chats:
    get:
      consumes:
      - application/json
      description: Retrieves all chats that the authenticated user is a participant
        of.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved chats
          schema:
            $ref: '#/definitions/api.ChatsList'
      security:
      - BearerAuth: []
      summary: Get all chats for a user
      tags:
      - chats
    post:
      consumes:
      - application/json
      description: Creates a new chat between two specified users and returns the
        chat hash.
      parameters:
      - description: Chat creation details
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/api.CreateChatMessage'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created chat
          schema:
            $ref: '#/definitions/api.ChatCreateResponse'
      security:
      - BearerAuth: []
      summary: Create a new chat
      tags:
      - chats
  /chats/{chatHash}:
    get:
      description: Retrieves the details of a specific chat and its messages if the
        user has access to it.
      parameters:
      - description: Chat hash
        in: path
        name: chatHash
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved chat details
          schema:
            $ref: '#/definitions/api.Chat'
      security:
      - BearerAuth: []
      summary: Get chat details by hash
      tags:
      - chats
...
swagger: "2.0"